feat: done ui&api: login and register seller admin account ; ui&setup for seller_admin home; done api create/edit/delete menu

+++ app/static/js/logout.js
@@ -2,14 +2,19 @@ document.addEventListener("DOMContentLoaded", () => {
     const username = localStorage.getItem("username");
     const usernameEl = document.getElementById("username-display");
 
-    if (username) {
+    if (username && usernameEl) {
         usernameEl.textContent = username;
         usernameEl.classList.remove("d-none");
     }
-});
 
-document.getElementById("logoutBtn").addEventListener("click", function (e) {
-    e.preventDefault();
-    localStorage.clear();
-    window.location.href = "/login";
+    const btnLogout = document.getElementById("logoutBtn");
+    if (btnLogout) {
+
+        btnLogout.addEventListener("click", function (e) {
+            e.preventDefault();
+            localStorage.clear();
+            window.location.href = "/login";
+        });
+    }
+
 });
